# perf: optimize memory usage for client.List func