--- a/client/api.go
+++ b/client/api.go
@@ -153,16 +153,13 @@ func (a api) List(ctx context.Context, result interface{}) error {
 		if err != nil {
 			return err
 		}
-		if resultVal.IsNil() || resultVal.Cap() == 0 {
-			resultVal.Set(reflect.MakeSlice(resultVal.Type(), 0, len(rows)))
-		}
 	} else {
 		rows = tableCache.Rows()
-		// If given a null slice, fill it in the cache table completely, if not, just up to
-		// its capability.
-		if resultVal.IsNil() || resultVal.Cap() == 0 {
-			resultVal.Set(reflect.MakeSlice(resultVal.Type(), 0, tableCache.Len()))
-		}
+	}
+	// If given a null slice, fill it in the cache table completely, if not, just up to
+	// its capability.
+	if resultVal.IsNil() || resultVal.Cap() == 0 {
+		resultVal.Set(reflect.MakeSlice(resultVal.Type(), 0, len(rows)))
 	}
 	i := resultVal.Len()
 	maxCap := resultVal.Cap()